+++ RaytraceGroundUp/Chapter28.3/Chapter28.3/GlobalTrace.h
@@ -15,7 +15,7 @@ class GlobalTrace: public Tracer {
         trace_ray(const Ray ray, const int depth) const;
 
         virtual RGBColor
-        trace_ray(const Ray ray, double& tmin, const int depth) const;
+        trace_ray(const Ray ray, float& tmin, const int depth) const;
 };
 
 #endif // GLOBALTRACE_H
